i am kitten -gene

+++ circleshape.py
@@ -12,7 +12,7 @@ class CircleShape(pygame.sprite.Sprite):
 
         self.position: pygame.Vector2 = pygame.Vector2(x, y)
         self.velocity: pygame.Vector2 = pygame.Vector2(0, 0)
-        self.radius: int = radius
+        self.radius: float = radius
 
     def draw(self, screen):
         # sub-classes must override
